move use-store-locator to hooks folder
