+++ packages/template-retail-react-app/app/components/store-locator/form.test.jsx
@@ -9,10 +9,10 @@ import React from 'react'
 import {render, screen} from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import {StoreLocatorForm} from '@salesforce/retail-react-app/app/components/store-locator/form'
-import {useStoreLocator} from '@salesforce/retail-react-app/app/components/store-locator/use-store-locator'
+import {useStoreLocator} from '@salesforce/retail-react-app/app/hooks/use-store-locator'
 import {useGeolocation} from '@salesforce/retail-react-app/app/components/store-locator/use-geo-location'
 
-jest.mock('./use-store-locator', () => ({
+jest.mock('@salesforce/retail-react-app/app/hooks/use-store-locator', () => ({
     useStoreLocator: jest.fn()
 }))
