+++ packages/template-retail-react-app/app/components/store-locator/form.tsx
@@ -16,7 +16,7 @@ import {
     Input
 } from '@chakra-ui/react'
 import {useForm, Controller} from 'react-hook-form'
-import {useStoreLocator} from '@salesforce/retail-react-app/app/components/store-locator/use-store-locator'
+import {useStoreLocator} from '@salesforce/retail-react-app/app/hooks/use-store-locator'
 import {useGeolocation} from '@salesforce/retail-react-app/app/components/store-locator/use-geo-location'
 import type {StoreLocatorFormValues} from '@salesforce/retail-react-app/app/components/store-locator/types'
