+++ packages/template-retail-react-app/app/components/store-locator/list.tsx
@@ -8,7 +8,7 @@
 import React, {useEffect, useState} from 'react'
 import {Accordion, AccordionItem, Box, Button} from '@chakra-ui/react'
 import {StoreLocatorListItem} from '@salesforce/retail-react-app/app/components/store-locator/list-item'
-import {useStoreLocator} from '@salesforce/retail-react-app/app/components/store-locator/use-store-locator'
+import {useStoreLocator} from '@salesforce/retail-react-app/app/hooks/use-store-locator'
 import {Stores, Store} from '@salesforce/retail-react-app/app/components/store-locator/types'
 
 interface StoreLocatorListProps {
